Organize project structure of at_app_bundler
