+++ packages/at_app_bundler/bin/main.dart
@@ -1,4 +1,4 @@
-import 'command_runner.dart';
+import 'package:at_app_bundler/at_app_bundler.dart';
 
 void main(List<String> args) async {
   await AtTemplateCommandRunner().run(args);
