+++ packages/at_app_bundler/lib/at_app_bundler.dart
@@ -1,0 +1,4 @@
+library at_app_bundler;
+
+export 'src/command_runner.dart';
+export 'src/commands/commands.dart';
